feat: move "Google Calendar" connection to sidepane
